fix(transitions): make transitions compatible with prefers-reduced-motion in svelte 5
Apparently, the behavior of transitions in Svelte 5 has changed slightly. Previously, if the user preferred reduced motion (prefers-reduced-motion), the `tick` callback was always called for the initial and final states. In Svelte 5, however, this is no longer called if the `duration` of the transition is 0. As a result, elements that used affected transitions could no longer become visible/invisible. To restore compatibility with Svelte 5, if the duration is 0, the designated end state is set directly.